requirements.txt
telethon>=1.42.0
python-dotenv>=1.0.0
pysocks>=1.7.1
httpx>=0.27.0
